refactor: introduce new TS project reference serverUtils
