+++ adminSiteServer/adminRouter.tsx
@@ -5,7 +5,7 @@ import rateLimit from "express-rate-limit"
 import filenamify from "filenamify"
 import * as React from "react"
 import { getConnection } from "typeorm"
-import { expectInt, tryInt, renderToHtmlPage } from "./serverUtil"
+import { expectInt, tryInt, renderToHtmlPage } from "../serverUtils/serverUtil"
 import { logInWithCredentials, logOut } from "./authentication"
 import { LoginPage } from "./LoginPage"
 import { RegisterPage } from "./RegisterPage"
