+++ adminSiteServer/tsconfig.json
@@ -14,6 +14,7 @@
         { "path": "../grapher" },
         { "path": "../clientUtils" },
         { "path": "../settings" },
-        { "path": "../baker" }
+        { "path": "../baker" },
+        { "path": "../serverUtils" }
     ]
 }
