+++ baker/SiteBaker.tsx
@@ -38,7 +38,7 @@ import { Post } from "../db/model/Post"
 import { bakeCountries } from "../baker/countryProfiles"
 import { countries } from "../clientUtils/countries"
 import { execWrapper } from "../db/execWrapper"
-import { logErrorAndMaybeSendToSlack } from "./slackLog"
+import { logErrorAndMaybeSendToSlack } from "../serverUtils/slackLog"
 import { countryProfileSpecs } from "../site/countryProfileProjects"
 import { ExplorerAdminServer } from "../explorerAdminServer/ExplorerAdminServer"
 import { getRedirects } from "./redirects"
